+++ cookbooks/maxlab_postfix/metadata.rb
@@ -3,7 +3,7 @@ maintainer 'Maxwell Spangler'
 maintainer_email 'maxcode@maxwellspangler.com'
 license 'All Rights Reserved'
 description 'Deploy postfix email services on internal nodes'
-version '1.1.0'
+version '1.2.0'
 chef_version '>= 14.0'
 
 # The `issues_url` points to the location where issues for this cookbook are
@@ -20,3 +20,5 @@ chef_version '>= 14.0'
 
 supports 'redhat', '>= 7.0'
 supports 'centos', '>= 7.0'
+
+depends 'chef-vault'