+++ policyfiles/tk_maxlab_postfix_server.rb
@@ -8,8 +8,10 @@ run_list "recipe[maxlab_firewall::base_firewalld]",
          "recipe[maxlab_postfix::deploy]",
          "recipe[maxlab_firewall::update_firewalld]"
 
-cookbook 'maxlab_postfix', path: '../cookbooks/maxlab_postfix'
-cookbook 'maxlab_firewall',   path: '../cookbooks/maxlab_firewall'
+cookbook 'maxlab_postfix',  path: '../cookbooks/maxlab_postfix'
+cookbook 'maxlab_firewall', path: '../cookbooks/maxlab_firewall'
+
+cookbook 'chef-vault'
 
 default['config_firewall']['base_config'] = 'base_firewall.maxlab.dmz'